HH-218702 add DatabaseSwitcher interface
